Fixed some deployment issues
Specifically in the newer versions of laravel fks must be bigints rather than ints, the seeder needs to be adjusted and finally to resore the db I had to make another change

+++ database/migrations/2020_04_27_104851_create_fks.php
@@ -14,17 +14,17 @@ class CreateFks extends Migration
     public function up()
     {
         Schema::table('uni_classes', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
         });
         Schema::table('grades', function (Blueprint $table) {
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->integer('uni_class_id')->unsigned()->index();
+            $table->bigInteger('uni_class_id')->unsigned()->index();
             $table->foreign('uni_class_id')->references('id')->on('uni_classes')->onDelete('cascade');
         });
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('role_id')->unsigned()->index();
+            $table->bigInteger('role_id')->unsigned()->index();
             $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
         });
     }
