+++ database/migrations/2020_04_27_110222_create_class_user_pivot_table.php
@@ -13,9 +13,9 @@ class CreateClassUserPivotTable extends Migration
     public function up()
     {
         Schema::create('uni_class_user', function (Blueprint $table) {
-            $table->integer('uni_class_id')->unsigned()->index();
+            $table->bigInteger('uni_class_id')->unsigned()->index();
             $table->foreign('uni_class_id')->references('id')->on('uni_classes')->onDelete('cascade');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->primary(['uni_class_id', 'user_id']);
         });
